- 무지, 그리드, 데일리 속지를 변경된 BasePaper에 맞춰서 수정 및 이름 변경
